pyproject.toml
[project]
name = "agent-adk-chainlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "asyncpg>=0.31.0",
    "azure-ai-inference>=1.0.0b9",
    "azure-identity>=1.25.1",
    "azure-storage-file-datalake>=12.22.0",
    "chainlit>=2.9.2",
    "ddgs>=9.9.1",
    "elevenlabs>=2.24.0",
    "google-adk>=1.19.0",
    "google-genai>=1.52.0",
    "greenlet>=3.2.4",
    "numpy>=2.3.5",
    "pymupdf>=1.26.6",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]
